requirements.txt
flask
langchain
langchain-community
sentence-transformers
faiss-cpu
numpy
flask-sqlalchemy
python-dotenv
langchain-groq
langchain-huggingface

